+++ library/src/device/generator/stockham_gen.cpp
@@ -305,7 +305,7 @@ std::string stockham_variants(const std::string&      filename,
     }
     else if(specs.scheme == "CS_KERNEL_STOCKHAM_BLOCK_CC")
     {
-        StockhamKernelCC kernel(specs);
+        StockhamKernelCC kernel(specs, false);
         output += append_common_functions(kernel.generate_lds_to_reg_input_function(),
                                           kernel.generate_lds_from_reg_output_function(),
                                           {},