+++ library/src/device/generator/stockham_gen_cc.h
@@ -23,8 +23,10 @@
 
 struct StockhamKernelCC : public StockhamKernel
 {
-    explicit StockhamKernelCC(const StockhamGeneratorSpecs& specs)
+    explicit StockhamKernelCC(const StockhamGeneratorSpecs& specs,
+                              bool                          largeTwdBatchIsTransformCount)
         : StockhamKernel(specs)
+        , largeTwdBatchIsTransformCount(largeTwdBatchIsTransformCount)
     {
         large_twiddle_steps.decl_default = 3;
         large_twiddle_base.decl_default  = 8;
@@ -37,6 +39,7 @@ struct StockhamKernelCC : public StockhamKernel
     Variable apply_large_twiddle{"apply_large_twiddle", "bool"};
     Variable large_twiddle_steps{"large_twiddle_steps", "size_t"};
     Variable large_twiddle_base{"large_twiddle_base", "size_t"};
+    bool     largeTwdBatchIsTransformCount = false;
 
     //
     // arguments
@@ -430,7 +433,7 @@ struct StockhamKernelCC : public StockhamKernel
                              Parens{large_twd_lds},
                              Parens{large_twiddles}};
         args.push_back(which);
-        args.push_back(transform);
+        args.push_back(largeTwdBatchIsTransformCount ? batch : transform);
         return args;
     }